% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_create}
\alias{file_create}
\alias{file_resize}
\alias{file_resize_off}
\title{Create or resize files}
\usage{
file_create(file, size)

file_resize(file, size)

file_resize_off(file, size)
}
\arguments{
\item{file}{Path name.}

\item{size}{Size in bytes.}
}
\value{
Input \code{file}, invisibly.
}
\description{
Create or resize files
}
\examples{
tmp <- tempfile()
file_create(tmp, 10)
file.size(tmp)
file_resize(tmp, 5)
file.size(tmp)
file_resize_off(tmp, 10)
file.size(tmp)

}
