% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBM-attach.R
\name{reconstruct_if_old}
\alias{reconstruct_if_old}
\title{Reconstruct FBM from old version}
\usage{
reconstruct_if_old(
  X,
  msg1 = "FBM from an old version? Reconstructing..",
  msg2 = "You should `$save()` it again."
)
}
\arguments{
\item{X}{An FBM.}

\item{msg1}{Message before reconstructing.}

\item{msg2}{Message after reconstructing.}
}
\value{
New FBM object (or old one if no need for reconstruction).
}
\description{
Reconstruct FBM from old version
}
\examples{
rds <- system.file("testdata", "before_readonly.rds", package = "rmio")
big_attach(rds)
}
\keyword{internal}
